refactor: rename `createDefaultConfigs` to `createConfigFile`
