+++ packages/dashmate/src/config/DefaultConfigs.js
